+++ server/src/main/java/com/linbit/linstor/netcom/PeerOffline.java
@@ -161,6 +161,17 @@ public class PeerOffline implements Peer
     }
 
     @Override
+    public void setAllowReconnect(boolean ignoredAllowReconnectRef)
+    {
+    }
+
+    @Override
+    public boolean isAllowReconnect()
+    {
+        return false;
+    }
+
+    @Override
     public void closeConnection()
     {
     }