+++ server/src/main/java/com/linbit/linstor/netcom/PeerTask.java
@@ -151,6 +151,17 @@ public class PeerTask implements Peer
     }
 
     @Override
+    public void setAllowReconnect(boolean ignoredAllowReconnectRef)
+    {
+    }
+
+    @Override
+    public boolean isAllowReconnect()
+    {
+        return false;
+    }
+
+    @Override
     public void closeConnection()
     {
     }