+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed race condition during "linstor node reconnect" leading to reconnect-cycle
+
 ## [1.31.0] - 2025-04-08
 
 ### Changed